+++ client/src/pages/Space/forms/CreateEntryForm/CreateEntryForm.scss
